updatupdating settings

+++ personal-settings/.bin-personal/sync-bashrc-personal.sh
@@ -6,5 +6,10 @@
 # Define common rsync options
 RSYNC_OPTS="-avuP"
 
-rsync $RSYNC_OPTS ~/.bashrc-personal /home/brett/bashrc-personal/.bashrc-personal
+# Copy the .bashrc-personal file from the local machine to Google Drive
+if [ ! -f /home/brett/bashrc-personal/.bashrc-personal ]; then
+  rsync $RSYNC_OPTS ~/.bashrc-personal /home/brett/bashrc-personal/.bashrc-personal
+fi
+
+# Copy the .bashrc-personal file from Google Drive to the local machine
 rsync $RSYNC_OPTS /home/brett/bashrc-personal/.bashrc-personal ~/.bashrc-personal
